[Impeller] Cache entire render target and not just allocations. (#50990)
This may be required to fix https://github.com/flutter/flutter/issues/144116 , and in general the safety of the render pass caches.

Because the RenderPass / Framebuffer objects refer to specific attachments, the caches are placed on the resolve texture, which is "real". The stencil/depth and or MSAA textures may be different but that didn't seem to matter on many Android devices. It doesn't seem like its actually valid though, whether or not there is a validation error to catch it.

I can confirm this fixes https://github.com/flutter/flutter/issues/144116  locally.

This change moves the cache one level up to cover the properties of the RenderTarget we care about, and ensures that we always use the same attachment combinations.

Fixes https://github.com/flutter/flutter/issues/141750
Maybe https://github.com/flutter/flutter/issues/144255 ?